requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
constantly==15.1.0
cryptography==37.0.1
cycler==0.11.0
daphne==3.0.2
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
distlib==0.3.4
Django==3.2
django-channels==0.7.0
django-crispy-forms==1.14.0
django-haystack==3.1.1
django-plotly-dash==1.6.6
django-windows-tools==0.2.1
dpd-components==0.1.0
elasticsearch==5.5.3
filelock==3.6.0
Flask==2.1.2
Flask-Compress==1.12
fonttools==4.33.2
future==0.18.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
msgpack==1.0.3
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
platformdirs==2.5.1
plotly==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.13.2
virtualenvwrapper-win==1.2.7
Werkzeug==2.0.3
zipp==3.8.0
zope.interface==5.4.0
